+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * New: `jigoshop_report_widgets` action to add custom report boxes.
     * New: `jQuery.payment()` function to ease payment redirection.
     * New: Ability to always select "All of" in country dropdown.
+    * New: Replaced old ThickBox with WordPress Media Gallery.
     * Improved: "Edit Product Category" and "Edit Product Tag" admin bar links now works properly.
     * Fix: Invalid formatting of shipping dropdown.
     * Fix: Displaying multiple select fields.